+++ R/wrapper-anticlustering.R
@@ -353,14 +353,15 @@ anticlustering <- function(x, K, objective = "diversity", method = "exchange",
   input_validation_anticlustering(x, K, objective, method, preclustering, 
                                   categories, repetitions, standardize, cannot_link)
 
-  NUMBER_OF_ANTICLUSTERS <- length(table(initialize_clusters(N, K, NULL)))
-  
   x <- to_matrix(x)
   N <- nrow(x)
   # there is a reason why scaling happens here and below (because of ILP + kplus)
   if (!is_distance_matrix(x) && standardize == TRUE) {
     x <- scale(x)
   }
+  
+  NUMBER_OF_ANTICLUSTERS <- length(table(initialize_clusters(N, K, NULL)))
+  TARGET_GROUPS <- table(initialize_clusters(N, K, NULL))
 
   ## Exact method using ILP
   if (method == "ilp") {
@@ -405,14 +406,25 @@ anticlustering <- function(x, K, objective = "diversity", method = "exchange",
   
   # BILS by Brusco et al.:
   if (method == "brusco") {
-    if (objective == "diversity") {
-      weights <- c(0.5, 0.99, 0.999, 0.999999)
-      obj_fun <- diversity_objective_
-    } else if (objective == "dispersion") {
+    weights <- c(0.5, 0.99, 0.999, 0.999999)
+    obj_fun <- diversity_objective_
+    average_diversity <- FALSE
+    if (objective == "dispersion") {
       weights <- c(0.000001, 0.00001, 0.0001, 0.001, 0.01, 0.1)
       obj_fun <- dispersion_objective_
+    } else if (objective == "kplus") {
+      x <- kplus_moment_variables(x, 2)
+      objective <- "variance"
+    } 
+    if (objective == "variance") {
+      x <- convert_to_distances(x)^2
+      if (any(TARGET_GROUPS != TARGET_GROUPS[1])) {
+        average_diversity <- TRUE
+      }
     }
-    partitions <- as.matrix(bicriterion_anticlustering(x, K, repetitions, weights))
+    partitions <- as.matrix(
+      bicriterion_anticlustering(x, K, repetitions, weights, average_diversity = average_diversity)
+    )
     # get best partition wrt dispersion / diversity
     best_obj <- which.max(apply(partitions, 1, obj_fun, convert_to_distances(x)))
     return(partitions[best_obj, ])