+++ NEWS.md
@@ -3,8 +3,8 @@
 - Bug fix in `optimal_dispersion()`: Output element `$edges` no longer includes edges that were investigated in the last iteration of the algorithm (and which are not relevant for finding the optimal dispersion)
 - `bicriterion_anticlustering()` has new arguments: `dispersion_distances`, `average_diversity`, `init_partitions`
 - `anticlustering()` now has new `objective = "average-diversity"`
-- **TODO**
-  * add argument `cannot_link` to `anticlustering()` (for `method = "exchange", "local-maximum", "brusco"` and `objective = "diversity", "variance", "kplus"`)
+- added argument `cannot_link` to `anticlustering()` (for `method = "exchange", "local-maximum", "brusco" "ilp"`, and `objective = "diversity", "variance", "kplus"`)
+- `method = "brusco"` now works for `objective = "variance"` and `"kplus"`
 
 # anticlust 0.8.2